+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog] and this project adheres to
 [Keep a Changelog]: http://keepachangelog.com/
 [Semantic Versioning]: http://semver.org/
 
+- Move existing library modules to Test.Discover.Internal
+
+# 4.2.3 [2022-05-21]
+
 - Added `--search-dir DIR` option
 - Adds an `--in-place` flag to write the generated driver to the source file.